refactor: change `Search` component layer from `widgets` to `views`
